+++ src/main/java/R/VD/goomong/post/service/PostService.java
@@ -10,10 +10,7 @@ import R.VD.goomong.item.repository.ItemRepository;
 import R.VD.goomong.member.model.Member;
 import R.VD.goomong.member.repository.MemberRepository;
 import R.VD.goomong.post.dto.request.RequestPostDto;
-import R.VD.goomong.post.exception.AlreadyDeletePostException;
-import R.VD.goomong.post.exception.NotExistPostCategoryException;
-import R.VD.goomong.post.exception.NotExistPostCategoryInfoException;
-import R.VD.goomong.post.exception.NotExistPostException;
+import R.VD.goomong.post.exception.*;
 import R.VD.goomong.post.model.Post;
 import R.VD.goomong.post.model.PostCategory;
 import R.VD.goomong.post.repository.PostCategoryRepository;
@@ -48,19 +45,25 @@ public class PostService {
     public void savePost(RequestPostDto requestPostDto, MultipartFile[] postImages, MultipartFile[] postFiles) {
         Post entity = requestPostDto.toEntity();
 
-        Member writer = memberRepository.findById(requestPostDto.getMemberId()).orElseThrow(() -> new RuntimeException("찾을 수 없는 회원입니다."));
+        Member writer = memberRepository.findById(requestPostDto.getMemberId()).orElseThrow(() -> new RuntimeException("해당 id의 회원은 없습니다. id = " + requestPostDto.getMemberId()));
+        if (writer.getDelDate() != null) throw new RuntimeException("해당 id의 회원은 삭제된 회원입니다. id = " + writer.getId());
 
         String postType = entity.getPostType();
         if ((postType.equals("커뮤니티") || postType.equals("FAQ")) && requestPostDto.getPostCategoryId() == null)
             throw new NotExistPostCategoryInfoException("카테고리를 선택해주세요");
 
         Item item = null;
-        if (requestPostDto.getItemId() != null)
-            item = itemRepository.findById(requestPostDto.getItemId()).orElseThrow(() -> new NotFoundItem("해당 상품이 없습니다."));
+        if (requestPostDto.getItemId() != null) {
+            item = itemRepository.findById(requestPostDto.getItemId()).orElseThrow(() -> new NotFoundItem("해당 id의 상품이 없습니다. id = " + requestPostDto.getItemId()));
+            if (item.getDelDate() != null) throw new RuntimeException("해당 id의 상품은 삭제된 상품입니다. id = " + item.getId());
+        }
 
         PostCategory postCategory = null;
-        if (requestPostDto.getPostCategoryId() != null)
-            postCategory = postCategoryRepository.findById(requestPostDto.getPostCategoryId()).orElseThrow(() -> new NotExistPostCategoryException("해당 카테고리가 없습니다."));
+        if (requestPostDto.getPostCategoryId() != null) {
+            postCategory = postCategoryRepository.findById(requestPostDto.getPostCategoryId()).orElseThrow(() -> new NotExistPostCategoryException("해당 id의 카테고리가 없습니다. id = " + requestPostDto.getPostCategoryId()));
+            if (postCategory.getDelDate() != null)
+                throw new AlreadyDeletePostCategoryException("해당 id의 카테고리는 삭제된 카테고리입니다. id = " + postCategory.getId());
+        }
 
         List<Image> postImageList = entity.getImageList();
         if (postImages.length != 0) postImageList = imageService.saveImage(postImages);
@@ -81,11 +84,13 @@ public class PostService {
     // 게시글 수정
     public Post updatePost(Long postId, RequestPostDto requestPostDto, MultipartFile[] postImages, MultipartFile[] postFiles) {
         Post onePost = findOnePost(postId);
+        if (onePost.getDelDate() != null)
+            throw new AlreadyDeletePostException("해당 id의 게시글은 삭제된 게시글입니다. id = " + onePost.getId());
 
-        List<Image> postImageList = null;
+        List<Image> postImageList = onePost.getImageList();
         if (postImages.length != 0) postImageList = imageService.saveImage(postImages);
 
-        List<Files> postFileList = null;
+        List<Files> postFileList = onePost.getFileList();
         if (postFiles.length != 0) postFileList = filesService.saveFiles(postFiles);
 
         Post build = onePost.toBuilder()
@@ -94,15 +99,14 @@ public class PostService {
                 .imageList(postImageList)
                 .fileList(postFileList)
                 .build();
-        postRepository.save(build);
-
-        return build;
+        return postRepository.save(build);
     }
 
     // 게시글 소프트딜리트
     public void softDeletePost(Long postId) {
         Post onePost = findOnePost(postId);
-        if (onePost.getDelDate() != null) throw new AlreadyDeletePostException("이미 삭제된 게시글입니다.");
+        if (onePost.getDelDate() != null)
+            throw new AlreadyDeletePostException("해당 id의 게시글은 이미 삭제된 게시글입니다. id = " + postId);
 
         Post build = onePost.toBuilder()
                 .delDate(LocalDateTime.now())
@@ -113,22 +117,41 @@ public class PostService {
     // 게시글 완전 삭제
     public void deletePost(Long postId) {
         Post onePost = findOnePost(postId);
+        if (onePost.getDelDate() != null)
+            throw new AlreadyDeletePostException("해당 id의 게시글은 이미 삭제된 게시글입니다. id = " + postId);
         postRepository.delete(onePost);
+    }
+
+    // 삭제된 게시글 복구
+    public void unDeleted(Long postId) {
+        Post origin = postRepository.findById(postId).orElseThrow(() -> new NotExistPostException("해당 id의 게시글은 없습니다. id = " + postId));
+        if (origin.getDelDate() == null) throw new NotDeletedPostException("해당 id의 게시글은 삭제된 글이 아닙니다. id = " + postId);
+
+        Post build = origin.toBuilder()
+                .delDate(null)
+                .build();
+        postRepository.save(build);
     }
 
     // 조회수 증가
     public void increaseViewCount(Long postId) {
+        Post post = postRepository.findById(postId).orElseThrow(() -> new NotExistPostException("해당 id의 게시글을 찾을 수 없습니다. id = " + postId));
+        if (post.getDelDate() != null) throw new AlreadyDeletePostException("해당 id의 게시글은 삭제된 게시글입니다. id = " + postId);
         postRepository.increaseViewCount(postId);
     }
 
     // 좋아요 증가
     public void increaseLikeCount(Long postId) {
+        Post post = postRepository.findById(postId).orElseThrow(() -> new NotExistPostException("해당 id의 게시글을 찾을 수 없습니다. id = " + postId));
+        if (post.getDelDate() != null) throw new AlreadyDeletePostException("해당 id의 게시글은 삭제된 게시글입니다. id = " + postId);
         postRepository.increaseLikeCount(postId);
     }
 
     // 하나의 게시글 조회
     public Post findOnePost(Long postId) {
-        return postRepository.findById(postId).orElseThrow(() -> new NotExistPostException("해당 게시글을 찾을 수 없습니다."));
+        Post post = postRepository.findById(postId).orElseThrow(() -> new NotExistPostException("해당 id의 게시글을 찾을 수 없습니다. id = " + postId));
+        if (post.getDelDate() != null) throw new AlreadyDeletePostException("해당 id의 게시글은 삭제된 게시글입니다. id = " + postId);
+        return post;
     }
 
     // 삭제되지 않은 게시글 조회
